dx9: mipmaps. load/dump textures. vmu/crosshair display.
upload all texture mipmaps to gpu
dump textures. load custom textures
vmu and lightgun xhair display
fix dx9 half-pixel offset
fix 565_32 unpacker bug

+++ core/rend/dx9/d3d_shaders.cpp
@@ -264,7 +264,7 @@ const ComPtr<IDirect3DPixelShader9>& D3DShaders::getShader(bool pp_Texture, bool
 		bool pp_Offset, u32 pp_FogCtrl, bool pp_BumpMap, bool fog_clamping,
 		bool trilinear, bool palette, bool gouraud)
 {
-	u32 hash = pp_Texture
+	u32 hash = (int)pp_Texture
 			| (pp_UseAlpha << 1)
 			| (pp_IgnoreTexA << 2)
 			| (pp_ShadInstr << 3)
